run.py: make the help for --dont-auto-calibrate less confusing
argparse.ArgumentDefaultsHelpFormatter displays the default value of the
arguments. This is an issue when the action is store_false, as then the
displayed default value is reversed (by default, autocalibration is
done, thus displaying dont-auto-calibrate=True by default is
misleading).

Storing the value in no_auto_calibrate instead solves the issue.

Change-Id: I523dfcc1c1a809cac4eb44f74fa45b8c2feae5b2