Update `utils_stats.py` to print median, MAD, and MAD (% median).
This patch updates `utils_stats.py` to report a new pair of statistics
for comparison of results. The rationale behind this change is that
the median is a very robust statistic, unaffected by outliers and
differences in runtimes between individual benchmarks. The MAD stands
for median absolute deviation, and is the counterpart of the standard
deviation when working with a median rather than a mean. The MAD is
calculated as the median of the absolute differences between each
data point in a list and the median of that list. Again, it is very
robust to outliers, and makes no assumptions about the underlying
distribution of the data (i.e. runtimes).

Change-Id: I8fa257211d392b8488fbb69bf41150e0105a4b36