Add Wilcoxon and t-test statistics.
Currently, when we compare two runs of the compilation statistics
framework, we only report the relative difference of means between
runs. This is a poor indicator of statistical significance, since it
assumes an equal and low standard deviation (or variance) between
runs.

Therefore, this patch introduces the Wilcoxon and Student's t-test
statistical tests, which report a p-value of how confident we are that
the difference between two consecutive runs is *not* statistically
significant. The Wilcoxon test is non-parametric, assumes nothing
about the distribution of the runtime data, and complements well the
median/median absolute deviation non-parametric statistics. The
Student's t-test is standard in many domains (psychology, medicine,
etc.) and assumes data comes from a normal distribution. Both a
non-parametric, and a parametric test were added to reduce the
possibility of mis-interpretation of the results.

Change-Id: I2a581b320b46618e3409214ffab4a0295c96a703