Clean up the compilation statistics script.
The main change is that now the first check done by the script is
if a command-line parameter is equal to `boot.oat`. Previously, if
there had been a file named `boot.oat` in the current directory,
the script would have interpreted it as an APK file, and it would
have run the standard compilation process.

The other major change is that duplicate command-line parameters
are ignored. Further changes include fixing some comments and
messages that have become slightly incorrect after the latest
changes, or simply removing them if they are no longer relevant.

Change-Id: I3be5e309170a2b0358aad850d80d735b7208d759