Change the output OAT file name to avoid clashes between different ISAs.
This patch changes the output OAT file name used by the compilation
statistics script to include the target ISA. The purpose is to avoid
name clashes when the user runs the same workload with different
`--mode` parameters, and needs to keep the output OAT files for
further analysis. Furthermore, this change makes the behaviour for
APK files consistent with the one for boot.oat.

Change-Id: I9eb16fb0bebd8a9a9b4df7c056a78cfd77af4a69