Fix Benchmark json output for linux devices
For linux devices, there is some extra output from dex2oat when running the benchmarks.
The filtering we had in place was not taking this into account.
As a result, this output was interpreted as a valid benchmark result and
some erratic benchmark results were included in the json.
This patch changes the filtering to include only the lines which end with 'per
iteration'. These are the only lines which contain benchmark results and
which we are interested in.

Change-Id: I91f5898d08937b6eede077a8f6a037b5d67a8255