Fix KeepSameKeys function.
Check if both inputs are dictionaries, and print a warning if not.

Remove iteration on stats_type: the json/dictionary is flat, there is
no higher level stats type as this function was initially expecting.

Finally, since the operation is done in place, there is no need to
return the dictionaries again.

Change-Id: I1a8b4ca1c49523c1ab42cee0dff8faf931451ba9