Fix checkstyle wrappers.
For now, use the version that we supported in the past, 6.14.1.  If we
find it useful, we will upgrade in the future.

Rework the filename convention for downloading based on the checkstyle
version.

For the download condition, make sure we check for the existence of the
file, and that its length is not 0 (-s vs -f).

Remove code relating to prompting the user for downloading checkstyle.
(seems unnecessary, and the python wrapper ignores that anyway).

Make sure we fail when the download (wget) fails. This is picked up by
the python wrapper which throws an exception when we have an error.

Check the wrapper argument list length and exit with 0 immediately if no
argument has been provided (including an info message). This conforms
well to the expectation of the Python script to download the checkstyle
binary when the bash wrapper is called without arguments.  We DO want to
know when the download fails.

Finally, just call checkstyle directly at the end of the bash wrapper,
ensuring that we get the output and the return code.

Change-Id: I2db89b71aeaa15f3af3c5eecb0bd693425d27565