Fix run.py result calculation when using the --dont-auto-calibrate option.
tools/benchmarks/run.py currently only records results from output lines
that end with 'per iteration'. These do not appear when the benchmarks
are run with the --dont-auto-calibrate option as no bench runner is used.
Instead, allow any line that starts with 'benchmarks/' to be parsed.

This prevents a crash that happens when run.py tries to calculate
the geomean while running with --dont-auto-calibrate.

Test: tools/benchmarks/run.py --iterations 1 --dont-auto-calibrate

Change-Id: I179a7c1b4a98288757bebcc721f1d4f6bd7b05f1