Fix java.io.FileNotFoundException from jar
When a path to benchmarks/build.sh used to run it is not absolute, the
jar tool used by the script fails with java.io.FileNotFoundException.
The exception happens because a path passed to the jar tool is not
valid. It is created relative to the current working directory. As
the current working directory is changed when the jar tool is used, the
path becomes invalid.

The CL fixes the error by making paths passed to the jar tool absolute.

Test: `pwd`/benchmarks/build.sh -t ./benchmarks/benchmarks/algorithm/DeltaBlue.java
Test: ./benchmarks/build.sh -t ./benchmarks/benchmarks/algorithm/DeltaBlue.java
Test: scripts/benchmarks/benchmarks_run_target.sh --mode 64 --cpu little --iterations 1 benchmarks/algorithm/Linpack
Change-Id: I873a0228298de23845cc653e5ec35336e6a71daf