Make to use different javac for target and host builds
When benchmark Java sources are compiled for a target, they are compiled
with '-source 1.8 -target 1.8'. However bootclasspath is not set to
libraries of the target. This causes that the built classes depend on
libraries of the host and can work incorrectly.

This CL makes to use different javac for target and host builds:
* For target: the script art/tools/javac-helper.sh is used to
correctly invoke javac. The script sets a correct bootclasspath.
* For host: javac from PATH is used.

In addition, bench.apk is built only for target builds.

Test: scripts/benchmarks/benchmarks_run_target.sh --mode 64 --cpu big benchmarks/specjvm2008/compress/CompressBench
Test: scripts/benchmarks/benchmarks_run_target.sh --mode 64 --cpu little --iterations 1 benchmarks/algorithm/Linpack
Test: ./benchmarks/build.sh ./benchmarks/benchmarks/algorithm/DeltaBlue.java
Test: env PATH=./prebuilts/jdk/jdk9/linux-x86/bin:$PATH ./benchmarks/build.sh ./benchmarks/benchmarks/algorithm/DeltaBlue.java
Change-Id: I6dc848fdf78918473605d2b6fc5655c592d63585